FEAT: Friendship 스크린타임 duration 타입 변경

+++ src/main/java/dev/wetox/WetoxRESTful/friendship/FriendshipListResponse.java
@@ -12,7 +12,7 @@ public class FriendshipListResponse {
 
     private Long userId;
     private String nickname;
-    private Double totalDuration;
+    private Long totalDuration;
 
 
 }
